+++ src/prospectScraper/Maps/ExistingProspectRankingCsvMap.cs
@@ -11,7 +11,7 @@ namespace prospectScraper.Maps
             Map(m => m.playerName).Name("Player");
             Map(m => m.school).Name("School");
             Map(m => m.position1).Name("Position");
-            Map(m => m.height).Name("Height").ConvertUsing(map => Program.ConvertHeightToInches(map.GetField("Height"), map.GetField("Player")).ToString());
+            Map(m => m.height).Name("Height").ConvertUsing(map => Player.ConvertHeightToInches(map.GetField("Height"), map.GetField("Player")).ToString());
             Map(m => m.weight).Name("Weight");
             Map(m => m.collegeClass).Name("CollegeClass");
             Map(m => m.rankingDateString).Name("Date");
